Move reverse-z code to rhi/rhi_policy.h
