+++ projects/Cyseal/Cyseal.vcxproj.filters
@@ -297,6 +297,9 @@
     <ClInclude Include="src\render\renderer_options.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
+    <ClInclude Include="src\rhi\rhi_policy.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\memory\mem_alloc.cpp">
